z6: fix review-flagged robustness bugs; resolve two Journey/Shogun items
Address the remaining items from the z6 code review (TODO_remaining_bugs.md).

Memory-safety / crashes:
- arthur.cpp: clamp width_in_chars >= 1 before divide/modulo (SIGFPE).
- writetiff.cpp: reject width <= 0 and images taller than 65535 rows.
- extract_image_data.cpp: validate directory_entry_size is 8/14/16 so the
  reservation check matches the bytes the parse loop consumes (heap OOB read).
- journey.cpp: bound both copy loops in print_tag_route_to_str (STRING_BUFFER
  _SIZE is only 15); clamp name_length up front.
- entrypoints.cpp: guard the find_shogun_globals dereferences with start != -1
  (memory[start-8] read at negative index on a mid-chain miss).

Integer overflow in image decoders:
- draw_png.cpp: size_t/ptrdiff_t size math, null-check calloc, validate IHDR
  dimensions (<= 8192) up front; widen draw_indexed_png canvas_size to size_t*.
- draw_image.cpp: size_t required_size + INT_MAX guard + fully-above-canvas
  early return in draw_bitmap_on_bitmap and draw_rectangle_on_bitmap.
- draw_apple_2.cpp: pass the size_t pixel-count product matching the buffer.

Uninitialized reads on truncated/corrupt streams:
- decompress_vga.cpp: calloc so an early end code leaves a transparent tail.
- draw_png.cpp: value-init the inflate buffer, check zlib return codes, and
  render only the bytes zlib actually produced.

Logic:
- v6_shared.cpp: signed math in select_hint_by_mouse (unsigned underflow);
  pass the real written length to win_menuitem in display_soft.

Verified against real binaries in ~/Downloads/{Journeys,Shoguns}:
- Deleted the dead CHANGE-NAME patch in entrypoints.cpp. It was an author-
  intended "later versions" patch that length_to_search=1 left permanently
  dead; enabling it breaks the VoiceOver name-change popup, so it served no
  purpose.
- Confirmed the Shogun V-COLOR 0x0d fallback order { &bg, &fg } is correct
  (r278/r283 reach it and yield fg/bg consistent with r288's primary path);
  documented the verification.

draw_image.cpp extract_palette left as-is (both palette producers allocate
>= 48 bytes); documented the invariant.

+++ terps/bocfel/z6/draw_apple_2.cpp
@@ -145,5 +145,8 @@ static uint8_t *decompress_apple2(ImageStruct *image) {
 // palette-indexed result to a 32-bit RGBA pixmap ready for display.
 uint8_t *draw_apple2(ImageStruct *image) {
     uint8_t *result = decompress_apple2(image);
-    return deindex(result, image->width * image->height, image);
+    // Use the same size_t product decompress_apple2 sized its buffer with;
+    // an int product could overflow and disagree, making deindex read past
+    // the decompressed buffer.
+    return deindex(result, (size_t)image->width * (size_t)image->height, image);
 }
